improve code in navigation library
