+++ core/navigation/src/main/java/com/christopher_elias/navigation/transactions/TransactionAnimations.kt
@@ -1,0 +1,15 @@
+package com.christopher_elias.navigation.transactions
+
+/*
+ * Created by Christopher Elias on 29/05/2021
+ * christopher.mike.96@gmail.com
+ *
+ * Loop Ideas
+ * Lima, Peru.
+ */
+
+enum class TransactionAnimations {
+    RIGHT_TO_LEFT,
+    BOTTOM_TO_TOP,
+    NONE
+}
