+++ core/navigation/src/main/java/com/christopher_elias/navigation/transactions/extensions/FragmentTransactionExtension.kt
@@ -1,9 +1,16 @@
-package com.christopher_elias.navigation.transactions
+package com.christopher_elias.navigation.transactions.extensions
 
 import android.util.Log
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
 import com.christopher_elias.navigation.R
+import com.christopher_elias.navigation.transactions.TransactionAnimations
+import com.christopher_elias.navigation.transactions.TransactionAnimations.NONE
+import com.christopher_elias.navigation.transactions.TransactionAnimations.BOTTOM_TO_TOP
+import com.christopher_elias.navigation.transactions.TransactionAnimations.RIGHT_TO_LEFT
+import com.christopher_elias.navigation.transactions.TransactionType
+import com.christopher_elias.navigation.transactions.TransactionType.ADD_FRAGMENT
+import com.christopher_elias.navigation.transactions.TransactionType.REPLACE_FRAGMENT
 
 /*
  * Created by Christopher Elias on 29/05/2021
@@ -37,63 +44,31 @@ internal fun FragmentTransaction.fragmentTransactionExt(
         "transactionType: $transactionType | addToBackStack: $addToBackStack | containerId: $containerId | fragmentTag: $fragmentTag | transactionAnimations: $transactionAnimations"
     )
     when (transactionAnimations) {
-        TransactionAnimations.NONE -> {
-            /* None transition animation is applied */
-        }
-        TransactionAnimations.BOTTOM_TO_TOP -> {
+        BOTTOM_TO_TOP -> {
             setCustomAnimations(
                 R.anim.slide_in_from_bottom, R.anim.slide_out_from_top,
                 R.anim.slide_in_from_bottom, R.anim.slide_out_from_top
             )
         }
-        TransactionAnimations.RIGHT_TO_LEFT -> {
+        RIGHT_TO_LEFT -> {
             setCustomAnimations(
                 R.anim.slide_in_right, R.anim.slide_out_left,
                 R.anim.slide_in_left, R.anim.slide_out_right
             )
         }
+        NONE -> {
+            /* None transition animation is applied */
+        }
     }
 
     when (transactionType) {
-        TransactionType.ADD_FRAGMENT -> {
-            add(
-                containerId,
-                newFragment,
-                fragmentTag
-            ).also {
-                if (addToBackStack) {
-                    it.addToBackStack(fragmentTag)
-                }
-            }.commit()
-        }
-        TransactionType.REPLACE_FRAGMENT -> {
-            replace(
-                containerId,
-                newFragment,
-                fragmentTag
-            ).also {
-                if (addToBackStack) {
-                    it.addToBackStack(fragmentTag)
-                }
-            }.commit()
-        }
+        ADD_FRAGMENT -> add(containerId, newFragment, fragmentTag)
+        REPLACE_FRAGMENT -> replace(containerId, newFragment, fragmentTag)
     }
-}
 
-enum class TransactionAnimations {
-    RIGHT_TO_LEFT,
-    BOTTOM_TO_TOP,
-    NONE
-}
+    if (addToBackStack) {
+        addToBackStack(fragmentTag)
+    }
 
-enum class TransactionType {
-    /**
-     * Add fragment on top of the current fragment that call this method.
-     */
-    ADD_FRAGMENT,
-
-    /**
-     * Replace the current fragment for a new fragment that call this method.
-     */
-    REPLACE_FRAGMENT
+    commit()
 }
